requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.2.0